Auth more like we should.
