+++ app/components/navbar-component/component.js
@@ -5,24 +5,28 @@ export default Ember.Component.extend({
     tagName: 'header',
     classNames: ['navbar', 'navbar-inverse', 'navbar-static-top'],
     session: Ember.inject.service(),
-    gravatarSrc: '',
-    userName: '',
 
-    init() {
-        this._super(...arguments);
-        var userData = this.get('session.data.userData');
+    gravatarSrc: Ember.computed('session.data.authenticated.user', function() {
+        let userData = this.get('session.data.authenticated.user');
         if (userData) {
-            this.set('gravatarSrc', userData.gravatar + '&s=25');
-            this.set('userName', `${userData.first_name} ${userData.last_name}`);
+            return userData.gravatar + '&s=25';
         }
-    },
+    }),
+
+    userName: Ember.computed('session.data.authenticated.user', function() {
+        let userData = this.get('session.data.authenticated.user');
+        if (userData) {
+            return `${userData.first_name} ${userData.last_name}`;
+        }
+    }),
+
     actions: {
         login(){
-            window.location = `${ENV.apiUrl}/accounts/login/`;
+            this.get('session').authenticate('authenticator:share');
         },
+
         logout(){
             this.get('session').invalidate();
-            this.get('session').set('data.userData', null);
         }
     }
 });
